fix gpt_voits
