+++ funasr/html/static/main.js
