+++ funasr/html/static/recorder-core.js
