+++ funasr/html/static/wav.js
