+++ funasr/html/static/wsconnecter.js
